daily-assisted-journal-backend/requirements.txt
flask 
flask-cors
python-dotenv
Flask-SQLAlchemy
psycopg2-binary
pytest